+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Security` in case of vulnerabilities.
 
 
+## Unreleased
+
+### Changed
+- extract `SharedConfig` to share type between `UserConfig` `UserFileConfig` `ResolvedUserConfig`
+
 ## [0.8.2] - 2022-04-21
 
 ### Added